feat: align templates folder
