--- src/templates/featureIssue.ts
@@ -1,31 +1,0 @@
-export const featureIssueTemplate: string[] = [
-  'name: 💡 Feature',
-  'description: Story related feature',
-  'title: "[FEATURE] <title>"',
-  'labels: ["type: feature"]',
-  'body:',
-  '  - type: textarea',
-  '    attributes:',
-  '      label: Description',
-  '      description: A description of the feature.',
-  '    validations:',
-  '      required: true',
-  '',
-  '  - type: textarea',
-  '    attributes:',
-  '      label: Task List',
-  '      description: Describe the steps to fulfill the feature.',
-  '      value: |',
-  '        - [ ] My First Task',
-  '    validations:',
-  '      required: true',
-  '# Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\".',
-];
-
-/**
- * Returns the feature issue template lines.
- * @returns An array of strings representing the feature template.
- */
-export function getFeatureIssueTemplateLines(): string[] {
-  return featureIssueTemplate;
-}
