+++ src/templates/github/bugIssue.ts
@@ -1,0 +1,27 @@
+/**
+ * Returns the bug issue template lines.
+ * @returns An array of strings representing the bug template.
+ */
+export function getBugIssueTemplateLines(): string[] {
+  return [
+    'name: 🐞 Bug',
+    'description: File a bug/issue',
+    'title: \'[BUG] < title >\'',
+    'labels: [\'type: bug\']',
+    'body:',
+    '  - type: textarea',
+    '    attributes:',
+    '      label: Current behavior',
+    '      description: A description of what you\'re experiencing.',
+    '    validations:',
+    '      required: true',
+    '',
+    '  - type: textarea',
+    '    attributes:',
+    '      label: Expected behavior',
+    '      description: A description of what you expected to happen.',
+    '    validations:',
+    '      required: true',
+    '# Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\".',
+  ];
+}
