+++ src/templates/github/questionIssue.ts
@@ -1,0 +1,20 @@
+/**
+ * Returns the question issue template lines.
+ * @returns An array of strings representing the question template.
+ */
+export function getQuestionIssueTemplateLines(): string[] {
+  return [
+    'name: ❓ Question',
+    'description: Ask a question',
+    'title: "[QUESTION] <title>"',
+    'labels: ["type: question"]',
+    'body:',
+    '  - type: textarea',
+    '    attributes:',
+    '      label: Question',
+    '      description: What would you like to know? If you encounter unusual behavior or identified a missing feature, consider opening a bug report instead.',
+    '    validations:',
+    '      required: true',
+    '# Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\".',
+  ];
+}
