--- src/templates/pullRequest.ts
@@ -1,33 +1,0 @@
-/**
- * Returns the pull request template lines.
- * @returns An array of strings representing the pull request template.
- */
-export const pullRequestTemplate: string[] = [
-  '## Reviewers Checklist',
-  '',
-  'for complete review list refer to ABS Loop - Review Aspects',
-  '',
-  '### Organizational Section',
-  '',
-  '- [ ] PR is assigned to the according feature/bug',
-  '- [ ] Feature/bug is descriptive',
-  '- [ ] Feature/bug is assigned to a developer',
-  '',
-  '### Structure',
-  '',
-  '- [ ] Readability: Code is easy to understand, with meaningful names for variables, functions, and classes',
-  '- [ ] Comments: Meaningful and helpful comments. Code is documented without being over-commented',
-  '- [ ] DRY, KISS and YAGNI: Code implements only necessary features; no over-engineering',
-  '- [ ] No sensitive data (e.g., passwords, API keys) in the code',
-  '- [ ] No major updates for used packages',
-  '',
-  '### Testing',
-  '',
-  '- [ ] Code is locally tested by developer (if applicable)',
-  '- [ ] Automated tests pass successfully',
-  '<!-- Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\".-->',
-];
-
-export function getPullRequestTemplateLines(): string[] {
-  return pullRequestTemplate;
-}
